requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
Werkzeug==2.3.7
PyPDF2==3.0.1
ebooklib==0.18
python-magic==0.4.27
pdfminer.six==20221105
nltk==3.8.1
gunicorn==21.2.0
pytest==7.4.2 